Don't shallow copy default request context
Summary:
This is to be consistent with setContext, which doesn't call set/unset of the default context.
Main goal is to merge code in next diff.

I think the default context is never relevant, and should only be the current context in the internal of the async framework, in between executing coroutines.
We should probably enforce that request data is never set in the default context to guarantee correctness.

Reviewed By: djwatson

Differential Revision: D9119717

fbshipit-source-id: 8fa07520d506c02663f58f7616f1f390645997bc