use heterogeneous lookup and mutation in folly::dynamic
Summary:
This diff allows `folly::dynamic` to lookup and mutate using any type that is
convertible to `StringPiece` (i.e., `const char*`, `std::string`, `fbstring`)
without having to construct a `dynamic` object (which requires copying the
string).

Currently dynamic string types uses fnv32 hash, which is OK since F14 is going
to do a mixing step anyway; we are changing this in D8304491.

Reviewed By: yfeldblum

Differential Revision: D8299863

fbshipit-source-id: a154bfe8f17e1d83968cdc1755ec54b6eed4299e