Missed l-value calls to future::then in Benchmark.cpp
Summary:
Overall plan to modify Future<T>::then to be r-value qualified and use Future<T>::thenTry or Future<T>::thenValue. The goal is to disambiguate folly::Future and to improve type and lifetime safety of Future and its methods.

Add a std::move that was missed in the folly benchmark.

Reviewed By: shixiao

Differential Revision: D9197265

fbshipit-source-id: d02ab7b0dc22f6465910a80b2d0af1b54fb7bde7