Fix a bug in ElfFile::getSectionByName
Summary:
It first looks for the name, then looks for a section with that
offset. But strings can be overlaid; eg .rela.debug_info and .debug_info
can be overlapped, so we won't find .debug_info at all.

Instead, just iterate the sections checking to see if the name matches
the one we're looking for.

Reviewed By: ricklavoie, palmtenor

Differential Revision: D9809817

fbshipit-source-id: b165719c6f9004ecc964096677114bce357ece30