Jemalloc Huge Page Allocator
Summary:
An allocator that uses a jemalloc arena backed by 2MB huge pages.

This uses the jemalloc extent hooks to let jemalloc hand out memory backed by huge pages.
The huge pages are allocated when init(int nr_pages) is called, and does not grow (although extending it to grow in the future should be fairly straightforward).

madvise is used to mark the pages as huge. It does not guarantee success, and the only way to tell is by looking at /proc/<pid>/smaps and the total huge pages in /proc/<pid>/status.

The HugePageAllocator can be used as template parameter for stl and folly collections.

Reviewed By: yfeldblum

Differential Revision: D9297627

fbshipit-source-id: 734ca1b19972a8b5d7a12ee60cba1ffa3da50dc2