Deprecate legacy forms of Future::then
Summary:
Deprecate the legacy forms of Future::then to assist migration to thenValue and thenTry.

This is part of a longer process to improve type correctness for folly::Future. By separating the forms of then we no longer have ambiguity when passing a polymorphic lambda as a continuation and can more easily tweak the implementation to modernise it.

Reviewed By: yfeldblum, WillerZ

Differential Revision: D9974379

fbshipit-source-id: 26f986329cb5d2d33d366b8f9e7b2664656642ff