Change makeObserver to compute the first value inline
Summary:
Current thread was previously blocked waiting on the Future to be completed anyway, so there's no reason to not treat it like one of the ObserverManager threads (if it isn't one already) and run the computation inline.
Without this change nested makeObserver calls could result in a deadlock if recursion depth was greater than the number of ObserverManager threads.

Reviewed By: yfeldblum

Differential Revision: D10097363

fbshipit-source-id: eb781ed5aea0ad9415da40fbc7f1c15ed481e92f