RFC 6902 - JSON Patch representation
Summary: Consume `folly::dynamic` and if it matches JSON patch format, build a sequence of patch operations. Application is to be done separately, e.g. by a method of `folly::dynamic` by walking the operations vector (obtained after parsing the patch) and applying operations to the document. This mimics the approach we used previously with `folly::json_pointer`.

Reviewed By: yfeldblum

Differential Revision: D10017234

fbshipit-source-id: 1d0b9f0967e528b81abf7998c88abdaf70a2fce8