fix the Windows build
Summary:
It looks like I also broke the Windows build of folly in D10435249, since
FollyCompilerMSVC.cmake referenced the `folly_deps` target, which is now not
defined yet when `FollyCompilerMSVC.cmake` is included.

This changes the code so that `FollyCompilerMSVC.cmake` simply updates
`${FOLLY_LINK_LIBRARIES}`, which is used in a `target_link_libraries()` call
in `folly-deps.cmake`.  (I'm slightly surprised that CMake was happy before
with multiple separate `target_link_libraries()` calls for the `folly_deps`
target.)

Reviewed By: Orvid

Differential Revision: D10445483

fbshipit-source-id: 5277785bec3fd8527d01a098e130c945077703ee