folly/memory/Malloc.h: include a few more JEMalloc macro definitions
Summary:
`folly/memory/Malloc.h` is generally used as a shim for the JEMalloc headers, providing
weak declarations of JEMalloc functions so that they can be referenced in library code
that may or may not be linked with JEMalloc at build/run time.

This adds a few more macro definition stubs for when JEMalloc isn't available.

Reviewed By: yfeldblum

Differential Revision: D10395893

fbshipit-source-id: 7b45714297acd85afd3b222687feeda491fcb46d