Create a handle-type agnostic sockets API and back the Windows socket portability layer with it
Summary:
The end goal will be to have all calls to the sockets APIs go through the agnostic API rather than the current approach, as the current approach causes a limit of ~64k sockets, due to file descriptor limitations on Windows.
All this diff does for now is create a sockets API that is not backed by file descriptors on Windows.
This also shifts the sockets portability layer to be backed by these APIs until the sockets portability layer is removed.

Reviewed By: yfeldblum

Differential Revision: D10235797

fbshipit-source-id: 1c073231290da48832608f6fec272c3d675bdfd7