move TSAN annotation hook declarations out of header
Summary:
AnnotateBenignRaceSized is prototyped with a long size arg in
some third-party libraries and a size_t arg in others.  This diff moves
folly's prototypes for the TSAN hooks out of a widely-included header
file so that the incompatible declarations can coexist.  This function
is declared with C linkage, so the mismatch is not a problem at link time.

Reviewed By: yfeldblum

Differential Revision: D12881785

fbshipit-source-id: af1b7404d9c3d9b728105591c112143d7388f4f9