(folly/benchmark)(RFC) allow users to record customized counters during benckmark
Summary:
There could be cases that user want to record some of their own metrics during benchmark. This diff add another kinds of macro BENCHMARK_COUNTERS to make it happen, see Test plan for an example.

One problem is that it would be hard to output the result on the fly. As it needs to run all benchmarks first to know which customized counter are used, then it can output the proper column headers.

Reviewed By: yfeldblum

Differential Revision: D9874668

fbshipit-source-id: 2004bb4484e54c1edd65763d1727486c6905d884