folly/symbolizer: don't use ELF base address for executables
Summary:
I'm not entirely sure the original motivation for this, but for the current
executable, we've been normalizing addresses from the virtual address referenced
by the first "LOAD" program header instead of 0.  This appears to be a bug, as
it forms invalid addresses into the ELF file.

In the case of the signal handler test, this happens to work with gold, as the
initial section lookup just happens to hit in the `.debug_str` section (as opposed
to the expected `.text` section), and the following name translation then undoes
the base address normalization to get a correct lookup.

While this works for gold, it breaks for LLD and meant we had to opt-out folly
out (D13279459).

Reviewed By: yfeldblum

Differential Revision: D14119804

fbshipit-source-id: 53dcaeb72ed75e58f9a46b15799ce1af5ff80531