Add option to set decay time to 0 for huge page allocator
Summary:
We observed some RSS regression for certain use cases - setting decay time to 0 should fix this as freed regions will be released back to the kernel immediately instead of delayed.

The risk is that we lose regions before we can reallocate them and thus over time end up with holes in the huge page region.
If we see this becoming an issue for some use cases, we can add a parameter to control the decay time.

Releasing immediately should be OK for huge pages in most cases. For cases with frequent allocation and deallocation of large areas there can be a perf impact however.

Reviewed By: interwq

Differential Revision: D13853171

fbshipit-source-id: 83fb95ed9c9bcb6ebdd1f80d78347e5eec61c084